The GenFSM Module

GenFSM — Module

GenFSM module

Main module of the GenFSM package

Module Index

GenFSM.Verbosity

GenFSM.runsim

GenFSM.to_dict

GenFSM.@random_seed!

Detailed API

GenFSM.Verbosity — Type

primitive type Verbosity <: Enum{Int32} 32

Many models and functions accept a verbosity parameter.

Choose between: NONE, LOW, STD [default], HIGH and FULL.

GenFSM.runsim — Function

run(project,scenario;override)

Run a scenario of GenFSM

Parameters

project: project or region name

scenario: name of the scenario within the project

override: a dictionary (keys=>new value) which eventually overrides the settings as specified in the region/scenario (e.g. for output directories)

GenFSM.to_dict — Method

df_to_dict(df, dim_col, value_cols)
Convert a DataFrame to a dictionary specifying the column(s) to use  as keys and the columns to use as values.

GenFSM.@random_seed! — Macro

Seed the global random seed to a mix given an hash from the Module.function name and the global random seed. In this way each, the random flows in each function that call this macro at the beginnin remain unrelated ot the other ones, but provided the yeaml-read global_random_seed (and julia verison), the flow remains the same.