The Res Module (Forest Resources)

GenFSM.Res — Module

module Res

Forest resources module of GenFSM

Module Index

GenFSM.Res.Pixel

GenFSM.Res.init!!

GenFSM.Res.load_settings!

GenFSM.Res.make_pixels

Detailed API

GenFSM.Res.Pixel — Type

mutable struct Pixel

Pixel used in the Res module

Fields:

id::Tuple{Int64, Int64}: A (x,y) tuple. Note that x is longitude and y is latitude (from top left)

fvol::Matrix{Float64}: Forest volumes, Mm^3, by ft,dc

farea::Matrix{Float64}: Forest Area, ha, by ft,dc

ac::Vector{Float64}: Availability coefficient, ∈ [0,1] by ft

fixed_state::Vector{Float64}: Fixed pixel characteristics

Encoded characteristics of the pixel (soil, position, geomorphology...). These are assumed not to change in the timeframe of the simulation, even if we change FT that may have an effect on soil

var_state::Vector{Float64}: Time variable characteristics Encoded characteristics of the pixel that by scenario and/or time depends (climate...)

initialized::Bool: Initialized by at least one module TODO: maybe unneeded just set initialization_model to 0

initialization_model::Int64: Initialization model id. It defines the ML model for growth and mortality

GenFSM.Res.init!! — Method

init!!(pixels, settings, global_mask)

Delegate pixel initialization to each Res region.

Res.init!!() –> init!!(Val(regxx)),...) –> Resxx._init!!()

GenFSM.Res.load_settings! — Method

load_settings!(settings)

Load the YAML file for the Res module and the YAML files for each RES region.

GenFSM.Res.make_pixels — Method

make_pixels(nft,ndc,nx,ny)

Return an array of uninstantiated pixels.

Note that we follow here the geo standard, so pixels[2,1] is pixel x=2 (from left), y=1 (from nord)